requirements.txt
# LLM & Agent
groq>=0.9.0
langchain-groq>=0.1.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0

# Evaluation LLM (Gemini judge)
google-generativeai>=0.8.0

# Embeddings & Reranking
sentence-transformers>=3.0.0  # all-MiniLM-L6-v2 or BGE-small-en-v1.5
# cross-encoder: BGE-reranker-base or ms-marco-MiniLM-L-6-v2 (via sentence-transformers)

# Vector DBs
chromadb>=0.5.0       # local development
qdrant-client>=1.9.0  # production (Qdrant Cloud)

# BM25 keyword search
rank-bm25>=0.2.2

# RAG Evaluation
ragas>=0.1.0
datasets>=2.20.0

# PDF generation
fpdf2>=2.7.0

# UI
streamlit>=1.40.0

# Ingestion
pypdf>=4.0.0           # PDF text extraction (ISSN papers, dietary guidelines)
beautifulsoup4>=4.12.0 # NIH ODS HTML scraping

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
